Implement a Schnorr based signature scheme.
The scheme is based on EdDSA but use the sec256k1 curve instead of ed25519.

For the explaination, we will use x the private key and P = x*G the public key.

To sign, a scalar secret k is chosen and R = k*G is computed. R.y must be even, but this is easy to ensure by negating k if needed. Then e is computed via e = SHA256(R.x || P || m) and the scalar s is computed such as s = k - e*x . The signature is the couple (R.x, s).

To veryfy, one compute e = H(R.x || P || m) and verify that R = e*P + s*G by checking R.x is indeed what we expect and R.y is even.